+++ examples/src/Testing/index.js
@@ -1,0 +1,17 @@
+import ChildrenExcludesSelfTest from './ChildrenExcludesSelfTest'
+import CombineEventsWithTest from './CombineEventsWithTest'
+import InitialPropsTest from './InitialPropsTest'
+import PremountEventOrderTest from './PremountEventOrderTest'
+import TeardownExcludesIsolatedTest from './TeardownExcludesIsolatedTest'
+import TeardownTest from './TeardownTest'
+import UnmountErrorTest from './UnmountErrorTest'
+
+export {
+  ChildrenExcludesSelfTest,
+  CombineEventsWithTest,
+  InitialPropsTest,
+  PremountEventOrderTest,
+  TeardownExcludesIsolatedTest,
+  TeardownTest,
+  UnmountErrorTest
+}